requirements.txt
# Core Photonics & Layout
gdsfactory==8.16.0    # The main tool for generating the GDS
pydantic<2.10         # Pinning this often helps avoid GDSFactory version conflicts

# Vision & Image Processing
opencv-python         # For the grid overlay and basic vision
numpy                 # Required for image math
matplotlib            # For plotting debug graphs if needed
Pillow                # Required for Gemini image processing

# AI & API Interaction
openai                # If using GPT-4o
google-genai          # If using Gemini
pyyaml                # To parse the output from the AI

# Utilities
ipykernel             # Allows you to run code cells inside Cursor/VS Code
python-dotenv         # For loading environment variables from .env file